pyproject.toml
[tool.poetry]
name = "kb-assistance-backend"
version = "0.1.0"
description = ""
authors = ["LIM SHENG XUE <limshengxue0820@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.13,>=3.11.0"
llama-parse = "^0.6.21"
llama-index = "^0.12.34"
llama-cloud-services = "^0.6.21"
llama-index-vector-stores-chroma = "^0.4.1"
llama-index-embeddings-mistralai = "^0.3.0"
llama-index-storage-chat-store-postgres = "^0.2.1"
psycopg-binary = "^3.2.7"
fastapi = {extras = ["standard"], version = "^0.115.12"}
pandas = "^2.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
